Diversify snoozin recs and improve user interface
For "random" rec option, create my own random word selector using local text
dictionary. And factor in popularity to avoid music whose artist is
below a popularity threshold

Add "similar" rec option, that uses spotify's recommendation engine

Change organizer user interface to consist of a process of sending a
"musicrecs info"  email to the organizer, waiting for a response with a
link to a 'guesses' google form, and then sending an email to all the
participants including the guess form now.