Fix stylesheet 404 bug
Stylesheet was not being found when adding routes that had an
extra level (ex: 'round/<long_id>/submit'). This was because the
stylesheet was being referenced as a relative path from the current
location.

The solution was to use the 'url_for' function within the html to be
able to get the absolute url to the stylesheet that will work with any
route.